feat: implement high-precision dynamic crypto engine & enterprise security layer
[Core Architecture & Refactoring]
- Migrated to a professional Monorepo structure inside 'src/' directory for full-stack scalability.
- Implemented 'Clean Architecture' (Onion Architecture) to ensure strict dependency flow:
    - CryptoKeyLab.Domain: Centralized all interfaces, models, and immutable records.
    - CryptoKeyLab.Infrastructure: Implemented high-performance Data Access using Dapper.
    - CryptoKeyLab.Core: Developed central business rules and the magic Reflection Factory.
    - CryptoKeyLab.Cryptography: Isolated math-engine containing pure algorithm implementations.
    - CryptoKeyLab.API: Entry point for HTTP orchestration, Security, and API Docs.

[Security & The Gatekeeper]
- Implemented 'ApiKeyAuthFilter' (Action Filter): A zero-trust security wall enforcing X-API-KEY validation.
- Developed 'ApiKeyValidationService': Separated security business logic from the HTTP pipeline (SOLID/SRP).
- Key Storage Security: Integrated SHA-256 one-way hashing for secure API Key persistence in LocalDB.
- Resilience: Integrated .NET 9 'IExceptionHandler' for sanitized, production-ready Global Error Handling.

[Dynamic Hashing Engine & Factory Pattern]
- Reflection-Based Factory: Developed a 'HashFactory' that dynamically instantiates algorithms based on Database Metadata, allowing for "Plug-and-Play" algorithm expansion.
- Options Pattern: Introduced 'HashOptions' to support polymorphic inputs (e.g., HMAC requiring secret keys while SHA-256 remains parameterless).
- Server-Driven UI Logic: API now provides metadata (RequiresKey, IsSecure) to drive the MudBlazor frontend dynamically.

[Algorithms Implemented]
- SHA-256: Industry-standard cryptographic hashing.
- HMAC-SHA256: Secure keyed-hash message authentication.
- BLAKE3: Implemented using SIMD-accelerated 'Blake3' NuGet package for world-class speed.
- Whirlpool: Implemented using 'BouncyCastle' library for 512-bit high-security hashing.

[Performance & Benchmarking]
- Refactored 'CryptoResult' from long to double: Implemented high-precision benchmarking using 'Stopwatch.Elapsed.TotalMilliseconds' to track sub-millisecond execution times.

[Infrastructure & Packages]
- Integrated 'Dapper' & 'Microsoft.Data.SqlClient' for optimized Stored Procedure execution.
- Configured 'Scalar.AspNetCore' for modern, interactive API documentation.
- Created 'HashAlgorithms' metadata table with optimized indexing for runtime configuration.

CryptoKeyLab/
├── src/
│   ├── CryptoKeyLab.sln
│   ├── CryptoKeyLab.API/               (Presentation & Security)
│   │   ├── Filters/ApiKeyAuthFilter.cs
│   │   ├── Middleware/GlobalExceptionHandler.cs
│   │   ├── Controllers/HashController.cs
│   ├── CryptoKeyLab.Core/              (Business Orchestration)
│   │   ├── Factories/HashFactory.cs
│   │   ├── Services/ApiKeyValidationService.cs
│   ├── CryptoKeyLab.Domain/            (Contracts & Models)
│   │   ├── Interfaces/Cryptography/IHashAlgorithm.cs
│   │   ├── Models/Cryptography/CryptoResult.cs (High-precision)
│   ├── CryptoKeyLab.Infrastructure/    (Data Access Layer)
│   │   ├── Repositories/HashMetadataRepository.cs (Dapper/SP)
│   ├── CryptoKeyLab.Cryptography/      (The Math Engine)
│   │   ├── Hashing/Cryptographic/Sha256Algorithm.cs
│   │   ├── Hashing/Auth/HmacSha256Algorithm.cs
│   │   ├── Hashing/Cryptographic/Blake3Algorithm.cs
│   │   └── Hashing/Cryptographic/WhirlpoolAlgorithm.cs
└── README.md                           (Product Documentation)

NuGet Package Audit (Installed Today)
Blake3: High-performance SIMD hashing.
BouncyCastle.NetCore: For Whirlpool and advanced crypto.

+++ src/CryptoKeyLab.Domain/Models/Cryptography/CryptoResult.cs
@@ -9,6 +9,7 @@ namespace CryptoKeyLab.Domain.Models.Cryptography
     public record CryptoResult
     (
         string? OutPut,
-        long TimeTakenMilliSeconds
+        //long TimeTakenMilliSeconds
+        double TimeTakenMilliSeconds
     );
 }
